shopper_api/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
openai>=1.3.0
crewai>=0.1.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.24.0 